move save function to utils
